+++ aim/web/api/projects/views.py
@@ -1,7 +1,7 @@
 import os
 from datetime import timedelta
 from typing import Optional, Tuple
-
+from logging import getLogger
 from collections import Counter
 from fastapi import Depends, HTTPException, Query, Header
 from aim.web.api.utils import APIRouter  # wrapper for fastapi.APIRouter
@@ -21,6 +21,8 @@ from aim.storage.locking import AutoFileLock
 
 projects_router = APIRouter()
 
+logger = getLogger()
+
 
 @projects_router.get('/', response_model=ProjectApiOut)
 async def project_api():
@@ -29,11 +31,27 @@ async def project_api():
     if not project.exists():
         raise HTTPException(status_code=404)
 
+    # check if the index db was corrupted and deleted
+    index_corrupted = project.repo.is_index_corrupted()
+    if index_corrupted:
+        runs_corrupted = False  # prevent multiple alert banners in UI
+        logger.warning('Index db was corrupted and deleted. '
+                       'Please run `aim storage reindex` command to restore optimal performance.')
+    else:
+        # check are there any corrupted run chunks
+        runs_corrupted = len(project.repo.list_corrupted_runs()) > 0
+        if runs_corrupted:
+            logger.warning('Corrupted Runs were detected. '
+                           'Please run `aim runs rm --corrupted` command to remove corrupted runs. '
+                           'You can list corrupted run hashes using `aim runs ls --corrupted` command.')
+
     return {
         'name': project.name,
         'path': project.path,
         'description': project.description,
         'telemetry_enabled': 0,
+        'warn_index': index_corrupted,
+        'warn_runs': runs_corrupted,
     }