+++ aim/web/ui/src/types/services/models/projects/projectsModel.d.ts
@@ -4,6 +4,8 @@ export interface IProject {
   name?: string;
   path?: string;
   telemetry_enabled?: string | boolean;
+  warn_index?: boolean;
+  warn_runs?: boolean;
 }
 
 export interface IProjectParamsMetrics {