+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+### Enhancements:
+- Add feature to delete full experiments (mauricekraus)
+
+## 3.20.1 Jun 3, 2024
+
+### Enhancements:
+- Repurpose aim reindex command for index db recreation (mihran113)
+
+### Fixes
+- Handle index db corruption and warn in UI (mihran113) 
+- Handle and skip corrupted runs (alberttorosyan)
+
 ## 3.19.3 Apr 17, 2024
 - Resolve issue with new runs after tracking queue shutdown (mihran113)
 - Reset base path when opening new tabs (mihran113)